Improved secrets handling: fixed file update logic, enhanced tests, updated `pack.sh` script, and bumped version to 1.5.21.

+++ pack.sh
@@ -1,3 +1,7 @@
-rm -fr dist
-python3 -m build
-python3 -m twine upload  dist/*
+#!/bin/sh
+
+if [ -d ./dist ] ; then
+  rm -fr dist
+fi
+python -m build
+python -m twine upload  dist/*
